Added tests on test-utils.

+++ test-utils/src/test/java/net/havox/javatools/test/utils/random/ModelRandomGeneratorTest.java
@@ -140,7 +140,7 @@ public class ModelRandomGeneratorTest
   {
     double expectedMinValue = 0.0d;
     double expectedMaxValue = 1.0d;
-    double randomDouble = ModelRandomGenerator.randomFloat();
+    double randomDouble = ModelRandomGenerator.randomDouble();
 
     this.checkRandomDouble( randomDouble, expectedMinValue, expectedMaxValue );
   }
